app/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.0
pydantic>=2.8.0
google-cloud-storage==2.13.0
llama-index-core==0.12.22
llama-index-llms-ollama==0.5.3
llama-index-embeddings-huggingface==0.5.2
neo4j==5.16.0
llama-index-graph-stores-neo4j==0.4.6
llama-index-readers-file==0.4.3

